rename files named channel.ts to repeater.ts
